+++ tests/models/test_admission.py
@@ -23,9 +23,14 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
+import random
+from unittest.mock import patch
+
 from django.test import TestCase
+from django.utils.translation import gettext as _
 
 from continuing_education.models import admission
+from continuing_education.models.enums import admission_state_choices
 from continuing_education.tests.factories.admission import AdmissionFactory
 from continuing_education.tests.factories.person import ContinuingEducationPersonFactory
 
@@ -33,6 +38,7 @@ from continuing_education.tests.factories.person import ContinuingEducationPerso
 class TestAdmission(TestCase):
     def setUp(self):
         self.admission = AdmissionFactory()
+        self.submitted_admission = AdmissionFactory(state=admission_state_choices.SUBMITTED)
         self.person = ContinuingEducationPersonFactory()
 
     def test_search(self):
@@ -42,3 +48,31 @@ class TestAdmission(TestCase):
 
         nonexistent_admission = admission.search(person=self.person)
         self.assertFalse(nonexistent_admission.exists())
+
+    @patch('osis_common.messaging.send_message.send_messages')
+    def test_mail_sent_on_admission_state_changed(self, mock):
+        state_values = [x[0] for x in admission_state_choices.ADMIN_STATE_CHOICES]
+        state = random.choice(state_values)
+        self.submitted_admission.state = state
+        self.submitted_admission.save()
+        self.assertTrue(mock.called)
+        message_content = mock.call_args[0][0]
+        self.assertIn(_(self.submitted_admission.state), str(message_content['template_base_data']))
+        self.assertIn(self.submitted_admission.person_information.person.user.email, str(message_content['receivers']))
+
+    @patch('osis_common.messaging.send_message.send_messages')
+    def test_mail_not_sent_on_same_admission_state(self, mock):
+        self.submitted_admission.save()
+        self.assertFalse(mock.called)
+
+    @patch('continuing_education.business.admission._get_continuing_education_managers')
+    @patch('osis_common.messaging.send_message.send_messages')
+    def test_mail_sent_on_admission_submitted(self, mock_send, mock_managers):
+        self.admission.state = admission_state_choices.DRAFT
+        self.admission.save()
+        self.admission.state = admission_state_choices.SUBMITTED
+        self.admission.save()
+        self.assertTrue(mock_send.called)
+        self.assertTrue(mock_managers.called)
+        message_content = mock_send.call_args[0][0]
+        self.assertIn(_(self.admission.state), str(message_content['template_base_data']))
